Completed the struct to class refactor.

+++ src/point.h
@@ -1,6 +1,6 @@
 ﻿#pragma once
 
 struct Point {
-	int x{ 0 };
-	int y{ 0 };
+	double x{ 0 };
+	double y{ 0 };
 };
